+++ R/PlotProteinOverlap.R
@@ -36,15 +36,14 @@ PlotProteinOverlap <- function(MQCombined) {
         group_by(samples) %>%
         summarise(value = n())
 
-    ggplot(df_bin_stat, aes(x = "all", y = value,
-                            fill = as.character(samples))) +
+
+    getPalette <- colorRampPalette(brewer.pal(9, 'Blues'))
+
+    ggplot(df_bin_stat, aes(x = "all", y = value,fill = factor(samples))) +
         geom_col(col = "white", width = 0.3) +
-        scale_fill_brewer(palette = 1) +
+        scale_fill_manual(values = getPalette(nrow(df_bin_stat))) +
         ylab("Number of Proteins") +
         theme_bw() +
         ggtitle("Protein Overlap Between samples") +
-        labs(
-            fill = "samples",
-            x = NULL
-        )
+        labs(fill = "samples",x = NULL)
 }